pr comments; also cleaned up bullet record tests
